Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/report.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/record-header.py?file=report-same-site
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/record-header.py?file=report-cross-site
Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/record-header.py?file=report-same-site
Blocked access to external URL https://www.127.0.0.1:9443/fetch/metadata/resources/record-header.py?file=report-cross-site

Harness Error (FAIL), message = Unhandled rejection: assert_not_equals: got disallowed value "No header has been recorded"

PASS same-origin report
TIMEOUT same-site report Test timed out
TIMEOUT cross-site report Test timed out
FAIL same-origin report: sec-fetch-dest assert_equals: expected "report" but got ""
FAIL same-origin report: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL same-origin report: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS same-origin report: sec-fetch-user

Harness Error (FAIL), message = Unhandled rejection: assert_not_equals: got disallowed value "No header has been recorded"

PASS same-origin report
TIMEOUT same-site report Test timed out
TIMEOUT cross-site report Test timed out
FAIL same-origin report: sec-fetch-dest assert_equals: expected "report" but got ""
FAIL same-origin report: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL same-origin report: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS same-origin report: sec-fetch-user
